+++ src/pages/404/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+const Invalid:React.FC = () => {
+  return (
+    <div>
+        404
+    </div>
+  )
+}
+export default Invalid;
