--- src/pages/Invalid/Invalid.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-const Invalid:React.FC = () => {
-  return (
-    <div>
-        404
-    </div>
-  )
-}
-export default Invalid;
