--- src/pages/Login/Login.css
@@ -1,17 +1,0 @@
-.loginForm{
-    width: 400px;
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%,-50%);
-}
-#root{
-    background-image: url(./background.jpg);
-    background-size:100%,100%;
-}
-.title{
-    display: block;
-    text-align: center;
-    font-size: 20px;
-    margin-bottom: 20px;
-}
