+++ src/pages/Login/index.css
@@ -1,0 +1,17 @@
+.loginForm{
+    width: 400px;
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%,-50%);
+}
+#root{
+    background-image: url(./background.jpg);
+    background-size:100%,100%;
+}
+.title{
+    display: block;
+    text-align: center;
+    font-size: 20px;
+    margin-bottom: 20px;
+}
