--- src/pages/Register/Register.css
@@ -1,7 +1,0 @@
-.registerForm{
-    width: 500px;
-    position: fixed;
-    top: 50%;
-    left: 45%;
-    transform: translate(-50%,-50%);
-}
