+++ src/pages/Register/index.css
@@ -1,0 +1,7 @@
+.registerForm{
+    width: 500px;
+    position: fixed;
+    top: 50%;
+    left: 45%;
+    transform: translate(-50%,-50%);
+}
