--- src/pages/SandBox/About/About.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-const About: React.FC = () => {
-    return (
-        <div>
-            11
-        </div>
-    )
-}
-export default About;
