+++ src/pages/SandBox/About/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+const About: React.FC = () => {
+    return (
+        <div>
+            11
+        </div>
+    )
+}
+export default About;
