--- src/services/SandBox/SandBox.tsx
@@ -1,1 +1,0 @@
-export default 1;
